backend/requirements.txt
# Django Core
Django==4.2.26
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-yasg==1.21.11

# Database
psycopg2-binary==2.9.11

# Environment & Config
python-decouple==3.8
pytz==2025.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.8
pdfminer.six==20251107
pypdfium2==5.0.0
reportlab==4.0.7

# OCR
pytesseract==0.3.13
Pillow==11.3.0

# OpenAI Integration
openai==2.8.1

# HTTP & Networking
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2025.11.12
idna==3.11
sniffio==1.3.1
anyio==4.11.0

# Data Validation
pydantic==2.12.4
pydantic-core==2.41.5
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2
jiter==0.12.0

# Security & Cryptography
cryptography==46.0.3
PyJWT==2.10.1
cffi==2.0.0
pycparser==2.23

# Utilities
asgiref==3.11.0
sqlparse==0.5.3
packaging==25.0
inflection==0.5.1
uritemplate==4.2.0
PyYAML==6.0.3
tqdm==4.67.1
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.0